requirements.txt
chardet==3.0.4
csvImporter==0.1.3.2
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.8
django-adaptors==0.2.5
django-csvimport==2.11
django-import-export==1.2.0
django-json-response==1.1.5
djangorestframework==3.4.0
et-xmlfile==1.0.1
funcsigs==1.0.2
jdcal==1.4
mock==2.0.0
odfpy==1.4.0
openpyxl==2.6.0
pbr==5.1.2
python-dateutil==2.8.0
PyYAML==3.13
six==1.12.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.2.0
xlwt==1.3.0
